requirements.txt
fastapi>=0.63.0
jinja2
python-multipart
python-dotenv
requests
uvicorn
